Moved/tided up tests
