+++ src/DomainLib.Serialization.Json.Tests/DomainLib.Serialization.Json.Tests.csproj
@@ -1,0 +1,19 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net5.0</TargetFramework>
+    <OutputType>Library</OutputType>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.8.0" />
+    <PackageReference Include="NUnit" Version="3.12.0" />
+    <PackageReference Include="NUnit3TestAdapter" Version="3.17.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\DomainLib.Serialization.Json\DomainLib.Serialization.Json.csproj" />
+    <ProjectReference Include="..\DomainLib.Testing\DomainLib.Testing.csproj" />
+  </ItemGroup>
+
+</Project>
