+++ src/DomainLib.Tests/DomainLib.Tests.csproj
@@ -1,14 +1,14 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>netcoreapp3.1</TargetFramework>
+        <TargetFramework>net5.0</TargetFramework>
+        <OutputType>Library</OutputType>
     </PropertyGroup>
 
     <ItemGroup>
       <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.8.0" />
-      <PackageReference Include="NSubstitute" Version="4.2.2" />
       <PackageReference Include="NUnit" Version="3.12.0" />
-      <PackageReference Include="NUnit3TestAdapter" Version="3.16.1" />
+      <PackageReference Include="NUnit3TestAdapter" Version="3.17.0" />
     </ItemGroup>
 
     <ItemGroup>
